# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cloned WebApp API Services",
    "version": "1.0.0",
    "description": "Production-grade OpenAPI 3.1 specification generated by the AGY Website Cloner & Appification Plugin. Covers scraping, web crawling, schema extraction, API key management, workspace analytics, and customer inquiry endpoints.",
    "contact": {
      "name": "API Support Team",
      "url": "https://docs.firecrawl.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "tags": [
    {
      "name": "Core Scraping",
      "description": "Scraping, crawling, and live web mapping operations"
    },
    {
      "name": "AI Extraction",
      "description": "Structured JSON schema extraction and batch jobs"
    },
    {
      "name": "API Keys",
      "description": "Developer API key lifecycle and permissions"
    },
    {
      "name": "Workspace & Usage",
      "description": "Team usage metrics, credit tracking, and activity logs"
    },
    {
      "name": "Contact Inquiries",
      "description": "Customer message and quote requests"
    }
  ],
  "paths": {
    "/api/v1/scrape": {
      "post": {
        "tags": [
          "Core Scraping"
        ],
        "summary": "Scrape a Single Web Page",
        "description": "Turn any URL into clean, LLM-ready Markdown, HTML, raw HTML, screenshot, or links.",
        "operationId": "scrapeUrl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Page successfully scraped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid URL or parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawl": {
      "post": {
        "tags": [
          "Core Scraping"
        ],
        "summary": "Initiate a Domain Crawl Job",
        "description": "Asynchronously crawl all pages on a website up to a specified depth limit.",
        "operationId": "startCrawl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Crawl job initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid crawl configuration"
          }
        }
      }
    },
    "/api/v1/map": {
      "post": {
        "tags": [
          "Core Scraping"
        ],
        "summary": "Map Website URLs",
        "description": "Discover all indexed and reachable sub-URLs for a target domain in seconds without full HTML download.",
        "operationId": "mapDomain",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com"
                  },
                  "search": {
                    "type": "string",
                    "example": "blog"
                  },
                  "limit": {
                    "type": "integer",
                    "default": 100,
                    "example": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URLs mapped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "links": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "https://example.com/about",
                        "https://example.com/blog"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/extract": {
      "post": {
        "tags": [
          "AI Extraction"
        ],
        "summary": "Extract Structured Data with AI",
        "description": "Parse a webpage and map its contents into a custom JSON schema using AI instructions.",
        "operationId": "extractStructuredData",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data extracted matching schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List Workspace API Keys",
        "description": "Retrieve all active and revoked API keys for the authenticated workspace.",
        "operationId": "listApiKeys",
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "keys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiKey"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create a New API Key",
        "description": "Generate a new secret API key token for production or development.",
        "operationId": "createApiKey",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Production Backend"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "key": {
                      "$ref": "#/components/schemas/ApiKey"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usage": {
      "get": {
        "tags": [
          "Workspace & Usage"
        ],
        "summary": "Get Credit Consumption & Activity Logs",
        "description": "Returns real-time credit quota balance, active browser worker counts, and recent request logs.",
        "operationId": "getWorkspaceUsage",
        "responses": {
          "200": {
            "description": "Current workspace usage stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageStats"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contact": {
      "post": {
        "tags": [
          "Contact Inquiries"
        ],
        "summary": "Submit Customer Contact Form",
        "description": "Accept customer inquiries, quote requests, or service messages with email notification.",
        "operationId": "submitContactInquiry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inquiry received successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Your message has been sent successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error in submitted fields"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "ScrapeRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://firecrawl.dev"
          },
          "formats": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "markdown",
                "html",
                "rawHtml",
                "links",
                "screenshot"
              ]
            },
            "default": [
              "markdown"
            ],
            "example": [
              "markdown",
              "links"
            ]
          },
          "onlyMainContent": {
            "type": "boolean",
            "default": true
          },
          "waitFor": {
            "type": "integer",
            "default": 0,
            "description": "Milliseconds to wait before scraping"
          }
        }
      },
      "ScrapeResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "example": "Example Domain"
              },
              "markdown": {
                "type": "string",
                "example": "# Example Domain\n\nThis domain is for illustrative examples."
              },
              "metadata": {
                "type": "object",
                "properties": {
                  "statusCode": {
                    "type": "integer",
                    "example": 200
                  },
                  "creditsUsed": {
                    "type": "integer",
                    "example": 1
                  }
                }
              },
              "links": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "CrawlRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://docs.firecrawl.dev"
          },
          "limit": {
            "type": "integer",
            "default": 10,
            "example": 50
          },
          "maxDepth": {
            "type": "integer",
            "default": 2,
            "example": 3
          }
        }
      },
      "CrawlJobResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "string",
            "example": "job_c798b1a2-3f"
          },
          "status": {
            "type": "string",
            "enum": [
              "scraping",
              "completed",
              "failed"
            ],
            "example": "scraping"
          },
          "total": {
            "type": "integer",
            "example": 12
          },
          "creditsUsed": {
            "type": "integer",
            "example": 12
          }
        }
      },
      "ExtractRequest": {
        "type": "object",
        "required": [
          "url",
          "schema"
        ],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/pricing"
          },
          "prompt": {
            "type": "string",
            "example": "Extract all pricing tiers and feature items"
          },
          "schema": {
            "type": "object",
            "description": "JSON Schema object defining desired return shape"
          }
        }
      },
      "ExtractResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Structured output matching the requested JSON Schema"
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "key_98b7a6c5"
          },
          "name": {
            "type": "string",
            "example": "Production Key"
          },
          "token": {
            "type": "string",
            "example": "fc_live_79a8e2b9c04d81234f98a76e"
          },
          "prefix": {
            "type": "string",
            "example": "fc_live_79a8e2b9"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsed": {
            "type": "string"
          }
        }
      },
      "UsageStats": {
        "type": "object",
        "properties": {
          "creditsUsedThisMonth": {
            "type": "integer",
            "example": 309
          },
          "creditsTotal": {
            "type": "integer",
            "example": 500
          },
          "activeBrowsers": {
            "type": "integer",
            "example": 0
          },
          "maxBrowsers": {
            "type": "integer",
            "example": 2
          },
          "recentRequests": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "endpoint": {
                  "type": "string",
                  "example": "/v1/scrape"
                },
                "url": {
                  "type": "string",
                  "example": "https://firecrawl.dev"
                },
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "credits": {
                  "type": "integer",
                  "example": 1
                },
                "time": {
                  "type": "string",
                  "example": "2 mins ago"
                }
              }
            }
          }
        }
      },
      "ContactSubmission": {
        "type": "object",
        "required": [
          "name",
          "email",
          "message"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "subject": {
            "type": "string",
            "example": "Wood Edging Quote"
          },
          "message": {
            "type": "string",
            "example": "I need pricing for 200 oak veneer panels."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Invalid URL provided"
          },
          "code": {
            "type": "string",
            "example": "INVALID_REQUEST"
          }
        }
      }
    }
  }
}