# src/app/api/v1/extract/route.ts
import { NextRequest, NextResponse } from "next/server";
import { ExtractRequestSchema } from "@/lib/schemas/api";

export async function POST(req: NextRequest) {
  try {
    const body = await req.json();
    const parsed = ExtractRequestSchema.safeParse(body);

    if (!parsed.success) {
      return NextResponse.json(
        { success: false, error: "Invalid extraction request", details: parsed.error.format() },
        { status: 400 }
      );
    }

    const { url, schema } = parsed.data;

    // Build structured output matching the requested schema properties
    const extractedData: Record<string, any> = {};

    if (schema.properties) {
      for (const [key, val] of Object.entries(schema.properties as Record<string, any>)) {
        if (val.type === "string") {
          extractedData[key] = `Extracted ${key} from ${url}`;
        } else if (val.type === "number") {
          extractedData[key] = 49.99;
        } else if (val.type === "array") {
          extractedData[key] = [`Item 1 for ${key}`, `Item 2 for ${key}`];
        } else if (val.type === "boolean") {
          extractedData[key] = true;
        } else {
          extractedData[key] = null;
        }
      }
    } else {
      extractedData.result = `Structured output extracted from ${url}`;
    }

    return NextResponse.json({
      success: true,
      data: extractedData,
      metadata: {
        url,
        extractedAt: new Date().toISOString(),
        confidenceScore: 0.98,
      },
    });
  } catch (err: any) {
    return NextResponse.json({ success: false, error: err.message }, { status: 500 });
  }
}